+++ packages/dotmac-approvals/src/dotmac_approvals/manifest.py
@@ -23,7 +23,7 @@ from dotmac_approvals.models import PLATFORM_TABLES, TENANT_TABLES
 
 module = ModuleManifest(
     code="approvals",
-    version="0.1.0a5",
+    version="0.1.0a5+dev",
     core=False,
     short_code="approvals",
     migration_prefix="ap",